# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgroConsult API documentation",
        "description": "AgroConsult API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/reaction": {
            "post": {
                "tags": [
                    "Reaction"
                ],
                "summary": "Create new reaction",
                "description": "Create reaction",
                "operationId": "f352a78b09ea81e738728ec2edefe9c0",
                "requestBody": {
                    "description": "Create new reaction",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment_id",
                                    "profile_id",
                                    "reaction"
                                ],
                                "properties": {
                                    "comment_id": {
                                        "type": "number",
                                        "example": "3"
                                    },
                                    "profile_id": {
                                        "type": "number",
                                        "example": "3"
                                    },
                                    "reaction": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, filled input. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reaction/{reaction}": {
            "delete": {
                "tags": [
                    "Reaction"
                ],
                "summary": "delete reaction",
                "description": "delete ",
                "operationId": "87ca91a2a550c9a83d336a086a57bb65",
                "parameters": [
                    {
                        "name": "reaction",
                        "in": "path",
                        "description": "reaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/payment": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Payment list",
                "description": "index",
                "operationId": "PaymentIndex",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status to filter by",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create new payment",
                "description": "Create by comoany or recruter",
                "operationId": "5085569296cc62d45a938b8e363c7b84",
                "requestBody": {
                    "description": "Create new payment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "application_id",
                                    "amount"
                                ],
                                "properties": {
                                    "application_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "amount": {
                                        "type": "string",
                                        "example": "250000"
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "example": "payme"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, filled input. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}